sql/catalog/multiregion: move multiregion validation to its package
This code and these invariants are seen as the responsibility of the
sql-experience team, so, let's move them to a package that that team owns.

Release justification: Low risk code movement.

Release note: None
