+++ pkg/sql/catalog/multiregion/BUILD.bazel
@@ -2,13 +2,18 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "multiregion",
-    srcs = ["region_config.go"],
+    srcs = [
+        "region_config.go",
+        "validate_table.go",
+    ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/catalog/multiregion",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/sql/catalog",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
+        "//pkg/sql/sem/tree",
         "@com_github_cockroachdb_errors//:errors",
     ],
 )
