+++ pkg/sql/catalog/tabledesc/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
         "//pkg/sql/catalog/catprivilege",
         "//pkg/sql/catalog/colinfo",
         "//pkg/sql/catalog/descpb",
+        "//pkg/sql/catalog/multiregion",
         "//pkg/sql/catalog/schemaexpr",
         "//pkg/sql/catalog/typedesc",
         "//pkg/sql/parser",
